requirements.txt
Django==1.8.19
djangorestframework==3.5.4
future==0.17.1
